[fw] fix library bug which cause READ_STORAGE to fail
there is a bug in the library which cause
`FRAM_MB85RC_I2C::readArray()` to fail.

datasheet indicates that when doing a read, the I2C address used
to specify the memory address and request for data should be the
same.
but in the code, the lib always request from 0x50, which cause it
to always read from 0b000XXXXXXXX.

the commit bf7076ceb03a7a542c8412b6e71257f354e199e5 in FRAM_MB85RC_I2C
fixes this bug.